Hack together some new mesh reduction algorithms
The previous algorithm implemented didn't work well. Instead implement
some algorithms tailored to what we're trying to reduce. For the roof/floor
objects generated by ASM, we completely remove the floor as it is not
visible. For walls, we identify all flat faces and simplify them into 2
triangles each. UVs have been removed because they're difficult to deal
with, instead just rely on the vertex colours which look fine. This also
helps improve performance a bit.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>