Separate walls into separate objects, workaround Unity bug
When rendering a top-down orthographic view of the map, walls don't need
to be rendered because they aren't visible. Separate them into a separate
object so they can be culled. Also workaround a Unity bug, CombineInstance
doesn't seem to prevent the mesh it's pointing to from being destroyed, so
make sure we keep a reference to it.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>